Set default values for non-nullable fields source_data and source_state when importing work items sources using the bulk import pattern.
This is a quick fix to solve the regression in production. Need to make this more robust so that it also handles the case when a non-null value is passed to this function in the input list. This will fail at present since we will have duplicate keys ni the dict being passed to the input. Need to handle this properly.

[PO-572]