[PP-463] BBA/Rewire: Introduce a new version of sync_work_item (sync_work_item_returning_multiple) and rewire usages of sync_work_item with sync_work_item_returning_multiple.
Target: rewire commands.insert_work_item to return multiple work items.

To fix this in an isolated fashion, we also updated the callers to handle the insert operation which now returns a list
in a different end to end path all the way to the message_listener.py

Also changed the contract between the issue handler and the message processors for this case so that it now returns a dict with organization key, work_items_source_key and the list of work items
instead of single work item.

We will move the update work flow to the same path next.