[PP-463] BBA/Rewire: Introduce a new version of sync_work_item (sync_work_item_returning_multiple) and rewire usages of sync_work_item with sync_work_item_returning_multiple.
Target: rewire callers commands.update_work_item to call sync_work_item_returning_multiple and delete commands.update_work_items

To fix this we had to decouple the move and update operations in the message handler so the update operations
returns multiple but the move handler returns single work items as before.

We will fix those in a subsequent pass.