[PP-463] BBA/Rewire: Introduce a new version of sync_work_item (sync_work_item_returning_multiple) and rewire usages of sync_work_item with sync_work_item_returning_multiple.
Target: single reference to sync_work_item in
issue_moved logic in jira_message_handler.py.

Replaced this with a call to sync_work_item_returning_multiple and used the 0th value as the result instead.

This is probably not quite precisely correct, but I cant think of a use case where it would not give the right answer, so leaving it at that.